requirements.txt
ultralytics
torch
opencv-python
pandas
numpy
psycopg2-binary
sqlalchemy
fastapi
uvicorn
python-multipart
pydantic
streamlit
requests
python-dotenv
matplotlib
scikit-learn
kagglehub
pytest
httpx
